Created interfaces directory and moved model directory to jpa
